Better output order for layout; add sample diagrams

+++ sample/scripts/Build-DependencyGraph.ps1
@@ -1,32 +1,51 @@
 #!/usr/bin/env pwsh
 
-# ------------------------------------------------------------------------------
-# Builds depenedency graph for CompanyXyz assemblies
+<#
+.SYNOPSIS
+    Builds .NET assembly dependency graph diagram and output to diagram formats
+
+.DESCRIPTION
+
+    For details on C4 see: https://c4model.com/
+
+.EXAMPLE
+
+    scripts/Build-DependencyGraph.ps1 -Verbose
+
+#>
+[CmdletBinding()]    
+param ()
 
 # Import library functions
 #Remove-Module DotnetDependencyGraph
 Import-Module ../DotnetDependencyGraph.psm1 -Force
 
-
-# 0. Copy DLLs to working folder
-
 $working = Join-Path $PSScriptRoot "../working"
 $data = Join-Path $working "data"
-Remove-Item $data -Recurse
+$csvFilePath = Join-Path $working "Dependencies-CompanyXyz.csv"
+
+
+Write-Verbose "Copying DLLs to working folder $working"
+
+if (Test-Path $data) {
+    Remove-Item $data -Recurse
+}
 $newDirectory = New-Item -ItemType Directory -Path $data
 
 Get-ChildItem (Join-Path $PSScriptRoot "..\src\CompanyXyz.DependencySample.Worker\bin\Debug") -Filter CompanyXyz*.dll -Recur -File | `
     ForEach-Object { Copy-Item $_.FullName $data -Force }
 
+
+ Write-Verbose "Get references from DLLs, expand data with attribute details, and save to $csvFilePath"
+
 # $references = ls src\CompanyXyz.DependencySample.Worker\bin\Debug\net8.0\CompanyXyz.* | Get-ReferencedAssemblies
 
-
-# 1. List files, get references, expand data with attribute details, save to CSV (because expanding takes time)
-
-# TODO Run this bit in a new PowerShell Session otherwise all the DLLs are loaded and locked in our current session so executing again will fail
 Get-ChildItem (Join-Path $data "CompanyXyz*.*") | Get-ReferencedAssemblies | Resolve-AssemblyReferences | Export-CSV (Join-Path $working "Dependencies-CompanyXyz.csv")
 
 #Get-Content 'Dependencies-Manual-CompanyXyz.csv' | Add-Content 'Working\Dependencies-CompanyXyz.csv'
+
+
+Write-Verbose "Load from saved CSV, filter, convert to diagram output"
 
 # 2. Load saved CSV, filter (e.g. direct, included only), convert to DOT using the color table, save to file
 # Array of name regex + color pairs. Checks are applied in order until first match.
@@ -47,6 +66,9 @@ Import-CSV (Join-Path $working "Dependencies-CompanyXyz.csv") `
     | Where-Object { -not ($_.Assembly -match 'Test|Migrat') } `
     | ConvertTo-PlantUml "Company XYZ - Dependency Sample component diagram" -NameColor $nameColor -AdditionalContent $plantUmlAdditionalContent `
     | Out-File (Join-Path $PSScriptRoot "../docs/dotnet-dependencies-CompanyXyz.puml")
+
+
+Write-Verbose "Load from saved CSV, filter, convert to C4 (in PlantUML), and save"
 
 $nameTag = @{ `
     '^CompanyXyz\.DependencySample\.Worker' = 'Worker'
